requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
pydantic==2.11.7
pydantic-settings==2.10.1
twilio==9.6.3
sarvamai==0.1.10
python-dotenv==1.0.1
python-multipart==0.0.20
nnmnkwii==0.1.3
scipy==1.15.3
requests==2.32.4
pywav==1.1
aiofiles==24.1.0
aiohttp==3.11.12
starlette==0.46.2
google-generativeai==0.8.3